Move MySQL AAD logic from flexible_server_custom_common to flexible_server_custom_mysql.py
